more than time service :^) (#5)
* added TimeService

* add todo

* wip

* wip

* wip

* some things added :^)

* wip

* quizService finished

* forgot an exception hihi

* chryste panie

* creating quiz with category, POSTs returns object

* nwm co dodalem ale dzialaja git quizy wiec wrzucam

* Json ignore to prevent infinite loop

* get me return user

* get me return user FIX

* rezultaty, pokoje i quizy wzglednie dzialaja, poprawie na dniach

* just working

* room in Results

* rooms refactor ready

* results refactor ready

* results refactor nearly ready (few things not clear)

* feat/RestExceptionHandler

* db port back to 5432

* quiz categories (#3)

* category service and a minor refactor

* wip

* get me return user

* get me return user FIX

* rezultaty, pokoje i quizy wzglednie dzialaja, poprawie na dniach

* just working

* room in Results

* rooms refactor ready

* results refactor ready

* results refactor nearly ready (few things not clear)

* feat/RestExceptionHandler

* db port back to 5432

* wip

* minor fixes

---------


* exception package refactor

---------

+++ src/main/java/app/Quiz/jwzpQuizappProject/service/UserService.java
@@ -1,26 +1,41 @@
 package app.Quiz.jwzpQuizappProject.service;
 
-import app.Quiz.jwzpQuizappProject.exceptions.UserAlreadyExistsException;
-import app.Quiz.jwzpQuizappProject.models.auth.RegisterRequest;
+import app.Quiz.jwzpQuizappProject.exceptions.auth.PermissionDeniedException;
+import app.Quiz.jwzpQuizappProject.exceptions.users.UserAlreadyExistsException;
+import app.Quiz.jwzpQuizappProject.exceptions.users.UserNotFoundException;
+import app.Quiz.jwzpQuizappProject.models.auth.RegisterDto;
 import app.Quiz.jwzpQuizappProject.models.users.SecurityUser;
+import app.Quiz.jwzpQuizappProject.models.users.UserDto;
 import app.Quiz.jwzpQuizappProject.models.users.UserModel;
+import app.Quiz.jwzpQuizappProject.models.users.UserStatus;
 import app.Quiz.jwzpQuizappProject.repositories.UserRepository;
-import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+
 @Service
 public class UserService implements UserDetailsService {
 
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
+    final TokenService tokenService;
 
-    public UserService(UserRepository userRepository, PasswordEncoder passwordEncoder) {
+    public UserService(UserRepository userRepository, PasswordEncoder passwordEncoder, TokenService tokenService) {
         this.userRepository = userRepository;
         this.passwordEncoder = passwordEncoder;
+        this.tokenService = tokenService;
+    }
+
+    private boolean validateUserDeleteUserAuthorities(UserModel executor, UserModel receiver){
+        return executor.isAdmin() || executor == receiver;
+    }
+
+    private boolean validateUserEditUserAuthorities(UserModel executor, UserModel receiver){
+        return executor.isAdmin() || executor == receiver;
     }
 
     @Override
@@ -31,13 +46,14 @@ public class UserService implements UserDetailsService {
                 .orElseThrow(() -> new UsernameNotFoundException("User with email " + email + " was not found."));
     }
 
-    public void saveUser(RegisterRequest registerRequest) throws UserAlreadyExistsException {
-        if (userRepository.existsByEmail(registerRequest.email())) {
-            throw new UserAlreadyExistsException("Account with email " + registerRequest.email() + " already exists.");
+    public void saveUser(RegisterDto registerDto) throws UserAlreadyExistsException {
+        if (userRepository.existsByEmail(registerDto.email())) {
+            throw new UserAlreadyExistsException("Account with email " + registerDto.email() + " already exists.");
         }
-        var user = new UserModel(registerRequest.name(), registerRequest.email(), passwordEncoder.encode(registerRequest.password()));
+        var user = new UserModel(registerDto.name(), registerDto.email(), passwordEncoder.encode(registerDto.password()));
         userRepository.save(user);
     }
+
     public UserModel getUser(String email) {
         var user = userRepository.findByEmail(email);
         if (user.isEmpty()) {
@@ -45,13 +61,52 @@ public class UserService implements UserDetailsService {
         }
         return user.get();
     }
-//TODO somehow decode JWT and check if the user in it matches the one in the request body that wants to change something
 
-//    public void updateUser() {
-//        userRepository.
-//    }
-    public void deleteUser(String email) {
+    public void deleteUser(String token, String email) throws PermissionDeniedException {
+        var executor = tokenService.getUserFromToken(token);
+        var receiver = getUser(email);
+
+        if(!validateUserDeleteUserAuthorities(executor,receiver )){
+            throw  new PermissionDeniedException(executor.getName() +  " does not have authorities to delete " + receiver.getName());
+        }
+
         userRepository.deleteByEmail(email);
     }
 
+    public void deactivateUser(String token, long id) throws PermissionDeniedException, UserNotFoundException {
+        var executor = tokenService.getUserFromToken(token);
+        var receiver = getUserById(id);
+
+        if(!validateUserDeleteUserAuthorities(executor,receiver )){
+            throw  new PermissionDeniedException(executor.getName() +  " does not have authorities to delete " + receiver.getName());
+        }
+
+        receiver.setStatus(UserStatus.DEACTIVATED);
+
+        userRepository.save(receiver);
+    }
+
+    public List<UserModel> getMultipleUsersByName(String userName){
+        return this.userRepository.findByName(userName);
+    }
+
+    public List<UserModel> getAllUsers(){
+        return this.userRepository.findAll();
+    }
+
+    public UserModel getUserById(long id) throws UserNotFoundException {
+        return this.userRepository.findById(id).orElseThrow(() -> new UserNotFoundException("User with id=" + id + " doesn't exist"));
+    }
+
+    public void updateUser(String token, UserDto user) throws UserNotFoundException, PermissionDeniedException {
+        var executor = tokenService.getUserFromToken(token);
+        var receiver = getUserById(user.userId());
+
+        if(!validateUserDeleteUserAuthorities(executor,receiver )){
+            throw  new PermissionDeniedException(executor.getName() +  " does not have authorities to update " + receiver.getName());
+        }
+
+        // TODO: not sure how to do this...
+
+    }
 }
